added ncube thing and gitignore

+++ timing_lib.h
@@ -1,4 +1,4 @@
 #include <time.h>
 #include "matrix.h"
 
-double time_function(multiplication_function, matrix*, matrix*, matrix*, int);
+float time_function(multiplication_function, matrix*, matrix*, matrix*, int);
